frontity: Add the tests and move packages for the HTML escaping
