+++ packages/core/src/config/webpack/externals.ts
@@ -5,6 +5,7 @@ export default ({ target }: { target: Target }): Configuration["externals"] => {
   const config: Configuration["externals"] =
     target !== "server"
       ? {
+          he: "{}",
           "node-fetch": "window.fetch",
           url: "{ URL: window.URL }"
         }
