+++ packages/frontity/types.ts
@@ -1,1 +1,5 @@
 export * from "@frontity/types";
+
+export interface Decode {
+  (text: string): string;
+}
