--- packages/html2react/src/libraries/decode/client.ts
@@ -1,9 +1,0 @@
-import { Decode } from "../../../types";
-
-const decode: Decode = text => {
-  const textArea = document.createElement("textarea");
-  textArea.innerHTML = text;
-  return textArea.value;
-};
-
-export default decode;
